FIX : UserRepository로 구독 사용자관련 쿼리 메소드 이전
