+++ lawmaking/src/main/java/com/everyones/lawmaking/repository/BillLikeRepository.java
@@ -1,13 +1,11 @@
 package com.everyones.lawmaking.repository;
 
 import com.everyones.lawmaking.domain.entity.BillLike;
-import com.everyones.lawmaking.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
@@ -17,8 +15,6 @@ public interface BillLikeRepository extends JpaRepository<BillLike, Long > {
             "where bl.bill.id = :billId AND bl.user.id = :userId")
     Optional<BillLike> findByUserIdAndBillId(@Param("userId") long userId, @Param("billId") String billId);
 
-    @Query("select bl.user from BillLike bl " +
-            "where bl.bill.id = :billId")
-    List<User> findAllByBillId(@Param("billId") String billId);
+
 
 }
